Dodanie komentarzy i widoku postów

+++ base/views.py
@@ -1,11 +1,13 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 # rooms = [
 #     {'id':1, 'name':'Python'},
@@ -14,8 +16,12 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username = username)
@@ -29,11 +35,29 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, 'Niepoprawna nazwa użytkownika lub hasło')
-    return render(request, 'login_register.html', {})
+    return render(request, 'login_register.html', {'page':page})
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method == 'POST': 
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Błąd podczas rejestracji')
+
+    return render(request, 'login_register.html', {'form':form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -50,9 +74,23 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all().order_by('-created')
 
-    return render(request, 'room.html', {'room':room})
+    participants = room.participants.all()
 
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+    
+        return redirect('room', pk=room.id)
+
+    return render(request, 'room.html', {'room':room, 'room_messages':room_messages, 'participants':participants})
+
+@login_required(login_url = '/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -60,11 +98,15 @@ def createRoom(request):
         if form.is_valid():
             form.save()
             return redirect('home')
-    return render(request, 'room_form.html', {'form':form})
+    return render(request, 'room_form.html', {'form':form, 'messages':messages})
 
+@login_required(login_url = '/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Nie jestes tworca pokoju!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -73,9 +115,27 @@ def updateRoom(request, pk):
             return redirect('home')
     return render(request, 'room_form.html', {'form':form})
 
+@login_required(login_url = '/login')
 def deleteRoom(request, pk):
+    room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Nie jestes tworca pokoju!')
+    
     room = Room.objects.get(id=pk)  
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'delete.html', {'obj':room})
+
+
+@login_required(login_url = '/login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('Nie jestes tworca pokoju!')
+    
+    room = Room.objects.get(id=pk)  
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'delete.html', {'obj':message})
